+++ frontend/src/metabase/admin/databases/containers/DatabaseEditApp.tsx
@@ -1,21 +1,25 @@
 import type { ComponentType } from "react";
 import { useState } from "react";
 import { connect } from "react-redux";
+import { push } from "react-router-redux";
+import type { Route } from "react-router";
 
 import { t } from "ttag";
 import _ from "underscore";
 import { updateIn } from "icepick";
 
 import { useMount } from "react-use";
-import type { Location } from "history";
+import type { Location, LocationDescriptor } from "history";
 import title from "metabase/hoc/Title";
 
 import Breadcrumbs from "metabase/components/Breadcrumbs";
 import Sidebar from "metabase/admin/databases/components/DatabaseEditApp/Sidebar/Sidebar";
 import { getUserIsAdmin } from "metabase/selectors/user";
+import { useCallbackEffect } from "metabase/hooks/use-callback-effect";
 
 import { getSetting } from "metabase/selectors/settings";
 
+import { LeaveConfirmationModal } from "metabase/components/LeaveConfirmationModal";
 import LoadingAndErrorWrapper from "metabase/components/LoadingAndErrorWrapper";
 import { DatabaseForm } from "metabase/databases/components/DatabaseForm";
 import ErrorBoundary from "metabase/ErrorBoundary";
@@ -26,7 +30,6 @@ import type {
   DatabaseId,
 } from "metabase-types/api";
 import type { State } from "metabase-types/store";
-import useBeforeUnload from "metabase/hooks/use-before-unload";
 import Database from "metabase-lib/metadata/Database";
 
 import { getEditingDatabase, getInitializeError } from "../selectors";
@@ -73,6 +76,8 @@ interface DatabaseEditAppProps {
   isAdmin: boolean;
   isModelPersistenceEnabled: boolean;
   initializeError?: DatabaseEditErrorType;
+  route: Route;
+  onChangeLocation: (location: LocationDescriptor) => void;
 }
 
 const mapStateToProps = (state: State) => {
@@ -97,6 +102,7 @@ const mapDispatchToProps = {
   discardSavedFieldValues,
   deleteDatabase,
   selectEngine,
+  onChangeLocation: push,
 };
 
 type DatabaseEditErrorType = {
@@ -124,6 +130,8 @@ function DatabaseEditApp(props: DatabaseEditAppProps) {
     initializeDatabase,
     params,
     saveDatabase,
+    route,
+    onChangeLocation,
   } = props;
 
   const editingExistingDatabase = database?.id != null;
@@ -131,7 +139,11 @@ function DatabaseEditApp(props: DatabaseEditAppProps) {
 
   const [isDirty, setIsDirty] = useState(false);
 
-  useBeforeUnload(isDirty);
+  /**
+   * Navigation is scheduled so that LeaveConfirmationModal's isEnabled
+   * prop has a chance to re-compute on re-render
+   */
+  const [isCallbackScheduled, scheduleCallback] = useCallbackEffect();
 
   useMount(async () => {
     await reset();
@@ -145,6 +157,12 @@ function DatabaseEditApp(props: DatabaseEditAppProps) {
   const handleSubmit = async (database: DatabaseData) => {
     try {
       await saveDatabase(database);
+
+      if (addingNewDatabase) {
+        scheduleCallback(() => {
+          onChangeLocation("/admin/databases?created=true");
+        });
+      }
     } catch (error) {
       throw getSubmitError(error as DatabaseEditErrorType);
     }
@@ -195,6 +213,11 @@ function DatabaseEditApp(props: DatabaseEditAppProps) {
           />
         )}
       </DatabaseEditMain>
+
+      <LeaveConfirmationModal
+        isEnabled={isDirty && !isCallbackScheduled}
+        route={route}
+      />
     </DatabaseEditRoot>
   );
 }
